# docs/swagger.yaml
definitions:
...
  body.CreateProfile:
    properties:
      email:
        example: test@gmail.com
        type: string
      gender:
        example: male
        type: string
      interestedField:
        example: develop
        type: string
      interestedFieldDetail:
        example:
        - web
        - server
        items:
          type: string
        type: array
    type: object
...
  body.UpdateProfile:
    properties:
      fileId:
        example: fileId
        type: string
      interestedField:
        example: develop
        type: string
      interestedFieldDetail:
        example:
        - web
        - server
        items:
          type: string
        type: array
    type: object
...
  model.Profile:
    properties:
      accountId:
        example: accountId
        type: string
      createdAt:
        example: "2019-12-23 12:27:37"
        type: string
      gender:
        example: male
        type: string
      id:
        example: profileId
        type: string
      imageUrl:
        example: profile.image_url.com
        type: string
      interestedField:
        example: develop
        type: string
      interestedFieldDetail:
        example:
        - web
        - server
        items:
          type: string
        type: array
      updatedAt:
        example: "2019-12-23 12:27:37"
        type: string
    type: object
...
info:
  contact: {}
  license: {}
paths:
...
  /profiles:
    get:
      consumes:
      - application/json
      parameters:
      - description: profileId list
        in: query
        items:
          type: string
        name: id
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Profile'
      security:
      - AccessToken: []
      tags:
      - Profiles
    post:
      consumes:
      - application/json
      description: create profile
      parameters:
      - description: Create Profile data
        in: body
        name: CreateProfile
        required: true
        schema:
          $ref: '#/definitions/body.CreateProfile'
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Profile'
      security:
      - AccessToken: []
      tags:
      - Profiles
    put:
      consumes:
      - application/json
      description: update profile
      parameters:
      - description: update profile data
        in: body
        name: UpdateProfile
        required: true
        schema:
          $ref: '#/definitions/body.UpdateProfile'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Profile'
      security:
      - AccessToken: []
      tags:
      - Profiles
  /profiles/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: profile id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Profile'
      tags:
      - Profiles
...
securityDefinitions:
  AccessToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"